Move virtual Sigset implementation to a separate module
Pull request #788 implemented the `Sigset` struct and its related items
including `Iter` and `IntoIter` in the
`yash_env::system::virtual::signal` module, which are re-exported by the
`yash_env::system::virtual` module. However, the `virtual` module also
re-exports other items that are not related to signal sets, such as
`Process` and `FileBody`. This can cause confusion for users since the
names `Iter` and `IntoIter` are quite generic and do not clearly
indicate that they are related to signal sets.

To improve the clarity of the code, this commit moves the `Sigset`
struct and its related items to a new `sigset` submodule under
`yash_env::system::virtual`. The new module is defined as a public
module, and the `virtual` module does not re-export the items from the
`sigset` module, so users will access them as
`yash_env::system::virtual::sigset::Sigset`,
`yash_env::system::virtual::sigset::Iter`, etc., which makes it clear
that they are related to signal sets in the virtual system.
